feat/add validation middleware
