+++ routes/index.js
@@ -1,31 +1,12 @@
 const router = require('express').Router();
-const { celebrate, Joi } = require('celebrate');
 const { createUser, login, logout } = require('../controllers/users');
 const auth = require('../middlewares/auth');
 const { notFoundPathErrorText } = require('../utils/constantsText');
 const NotFoundError = require('../errors/not-found-err');
+const { validateCreateUser, validateLoginUser } = require('../middlewares/validation');
 
-router.post(
-  '/signup',
-  celebrate({
-    body: Joi.object().keys({
-      email: Joi.string().required().email(),
-      password: Joi.string().required().min(8),
-      name: Joi.string().required().min(2).max(30),
-    }),
-  }),
-  createUser,
-);
-router.post(
-  '/signin',
-  celebrate({
-    body: Joi.object().keys({
-      email: Joi.string().required().email(),
-      password: Joi.string().required().min(8),
-    }),
-  }),
-  login,
-);
+router.post('/signup', validateCreateUser, createUser);
+router.post('/signin', validateLoginUser, login);
 router.post('/signout', logout);
 
 router.use('/users', auth, require('./users'));
