+++ routes/users.js
@@ -1,20 +1,11 @@
 const router = require('express').Router();
-const { celebrate, Joi } = require('celebrate');
 const {
   getCurrentUser,
   updateUserProfile,
 } = require('../controllers/users');
+const { validateUpdateUserProfile } = require('../middlewares/validation');
 
 router.get('/me', getCurrentUser);
-router.patch(
-  '/me',
-  celebrate({
-    body: Joi.object().keys({
-      name: Joi.string().required().min(2).max(30),
-      email: Joi.string().required().email(),
-    }),
-  }),
-  updateUserProfile,
-);
+router.patch('/me', validateUpdateUserProfile, updateUserProfile);
 
 module.exports = router;
